working setup progress bar

+++ src/com/mechinn/android/ouralliance/Setup.java
@@ -2,27 +2,38 @@ package com.mechinn.android.ouralliance;
 
 import java.io.File;
 import java.sql.SQLException;
+import java.util.HashMap;
+import java.util.Map;
 
 import android.app.Activity;
+import android.app.FragmentManager;
+import android.content.ContentResolver;
+import android.database.sqlite.SQLiteDatabase;
+import android.net.Uri;
+import android.os.AsyncTask;
+import android.os.Bundle;
 import android.os.Environment;
 import android.util.Log;
-import android.util.SparseArray;
 
 import com.mechinn.android.ouralliance.data.Competition;
 import com.mechinn.android.ouralliance.data.CompetitionTeam;
 import com.mechinn.android.ouralliance.data.Season;
-import com.mechinn.android.ouralliance.data.Team;
-import com.mechinn.android.ouralliance.data.frc2013.TeamScouting2013;
 import com.mechinn.android.ouralliance.data.source.CompetitionDataSource;
 import com.mechinn.android.ouralliance.data.source.CompetitionTeamDataSource;
 import com.mechinn.android.ouralliance.data.source.SeasonDataSource;
 import com.mechinn.android.ouralliance.data.source.TeamDataSource;
 import com.mechinn.android.ouralliance.data.source.frc2013.TeamScouting2013DataSource;
 import com.mechinn.android.ouralliance.error.OurAllianceException;
+import com.mechinn.android.ouralliance.provider.DataProvider;
+import com.mechinn.android.ouralliance.view.LoadingDialogFragment;
+import com.mechinn.android.ouralliance.view.TeamListFragment.Listener;
 
-public class Setup {
-	private static final String TAG = "Setup";
-	private static final int VERSION = 1;
+public class Setup extends AsyncTask<Void, Object, Boolean> {
+	public static final int PRIMARY = 148;
+	public static final int VERSION = 1;
+	public static final String TAG = "Setup";
+	private static final int INDETERMINATE = -1;
+	private static final int NORMAL = 0;
 	private Prefs prefs;
 	private TeamDataSource teamData;
 	private SeasonDataSource seasonData;
@@ -31,8 +42,30 @@ public class Setup {
 	private CompetitionTeamDataSource competitionTeamData;
 	private String packageName;
 	private File dbPath;
+	private Integer flag;
+	private Integer primary;
+	private Integer version;
+	private CharSequence status;
+	private Map<String, String> comps;
+	private LoadingDialogFragment dialog;
+	private ContentResolver data;
+	private boolean reset;
+	private FragmentManager fragmentManager;
+    private Listener listener;
+
+    public interface Listener {
+        public void setupComplete();
+    }
 	
-	public Setup(Activity activity) {
+	public Setup(Activity activity, boolean reset) {
+        try {
+        	listener = (Listener) activity;
+        } catch (ClassCastException e) {
+            throw new ClassCastException(activity.toString() + " must implement "+TAG+".Listener");
+        }
+		this.reset = reset;
+		this.fragmentManager = activity.getFragmentManager();
+		this.data = activity.getContentResolver();
 		prefs = new Prefs(activity);
 		teamData = new TeamDataSource(activity);
 		seasonData = new SeasonDataSource(activity);
@@ -43,17 +76,51 @@ public class Setup {
 		dbPath = activity.getDatabasePath("ourAlliance.sqlite");
 	}
 	
-	public void reset() {
-		prefs.clear();
-		run();
+	@Override
+	protected void onPreExecute() {
+		CharSequence title = "Setup data";
+		if(reset) {
+			prefs.clear();
+			prefs.setVersion(-1);
+			title = "Reset data";
+		}
+		Log.d(TAG,"curent version: "+prefs.getVersion());
+		Log.d(TAG,"new version: "+VERSION);
+		if(prefs.getVersion() < VERSION) {
+			dialog = new LoadingDialogFragment();
+			Bundle dialogArgs = new Bundle();
+			dialogArgs.putCharSequence(LoadingDialogFragment.TITLE, title);
+			dialogArgs.putInt(LoadingDialogFragment.MAX, PRIMARY);
+			dialog.setArguments(dialogArgs);
+            dialog.show(fragmentManager, "Setup Our Alliance");
+		} else {
+			this.cancel(true);
+		}
 	}
-	
-	public void run() {
-		int version = prefs.getVersion();
+
+	@Override
+	protected Boolean doInBackground(Void... params) {
+		version = prefs.getVersion();
 		Log.d(TAG, "version: "+version);
+		flag = 0;
+		primary = 0;
 		switch(version+1) {
+			case 0:
+				increaseVersion();
+//		        if(SQLiteDatabase.deleteDatabase(dbPath)) {
+//					Log.d(TAG,"deleted db");
+//				} else {
+//					Log.d(TAG,"did not delete db");
+//				}
+				setStatus("Resetting database");
+		        data.call(Uri.parse(DataProvider.BASE_URI_STRING+TAG), DataProvider.RESET, null, null);
+	        	setFlag(NORMAL);
 			case 1:
+				increaseVersion();
+	        	setFlag(INDETERMINATE);
 		        if(Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED)) {
+		        	setFlag(INDETERMINATE);
+					setStatus("Deleting old picture directory");
 			        File externalPath = Environment.getExternalStorageDirectory();
 			        File picDir = new File(externalPath.getAbsolutePath() +  "/Android/data/" + packageName + "/files");
 			        Utility.deleteRecursive(picDir);
@@ -62,141 +129,167 @@ public class Setup {
 			        	picDir.mkdirs();
 			        }
 		        }
-
-//				if(SQLiteDatabase.deleteDatabase(dbPath)) {
-//					Log.d(TAG,"deleted db");
-//				} else {
-//					Log.d(TAG,"did not delete db");
-//				}
-			default:
-		}
-		switch(version+1) {
-			case 1:
+				setStatus("Adding 2013 data");
 				try {
-					Season s2012 = new Season(2012, "Rebound Rumble");
-					seasonData.insert(s2012);
-					competitionData.insert(new Competition(s2012, "WORLD CHAMPS", "WORLD"));
 					Season s2013 = new Season(2013, "Ultimate Ascent");
+					Log.d(TAG,s2013.toString());
 					s2013 = seasonData.insert(s2013);
-					this.addCompetitions(s2013);
+		        	increasePrimary();
+					if(!addCompetitions(s2013)) {
+						return false;
+					}
 				} catch (OurAllianceException e) {
 					e.printStackTrace();
+					return false;
 				} catch (SQLException e) {
 					e.printStackTrace();
+					return false;
 				}
+//			case 2:
+//
+//				increaseVersion();
+		}
+		setStatus("Finished");
+		prefs.setVersion(version);
+		return true;
+	}
+	
+	@Override
+	protected void onProgressUpdate(Object... progress) {
+		int flag = (Integer)progress[0];
+		int primary = (Integer)progress[1];
+		dialog.setProgressStatus((String)progress[2]);
+		switch(flag) {
+			case INDETERMINATE:
+				dialog.setProgressIndeterminate();
+			case NORMAL:
 			default:
+				dialog.setProgressPercent(primary);
 		}
-		prefs.setVersion(VERSION);
-	}
+    }
+
+	@Override
+    protected void onPostExecute(Boolean result) {
+		dialog.dismiss();
+		listener.setupComplete();
+    }
 	
-	private void addCompetitions(Season season) {
-		Competition comp = new Competition(season, "Alamo Regional", "STX");
-		try {
-			comp = competitionData.insert(comp);
-			addTestData(season, comp);
-//			competitionData.insert(new Competition(season, "Alamo Regional", "STX"));
-			competitionData.insert(new Competition(season, "BAE Systems Granite State Regional", "NH"));
-			competitionData.insert(new Competition(season, "Greater Kansas City Regional", "KC"));
-			competitionData.insert(new Competition(season, "Hatboro-Horsham FIRST Robotics District Competition", "PAH"));
-			competitionData.insert(new Competition(season, "Kettering University FIRST Robotics District Competition", "GG"));
-			competitionData.insert(new Competition(season, "Gull Lake FIRST Robotics District Competition", "MIGL"));
-			competitionData.insert(new Competition(season, "Smoky Mountains Regional", "TN"));
-			competitionData.insert(new Competition(season, "San Diego Regional", "SDC"));
-			competitionData.insert(new Competition(season, "Israel Regional", "IS"));
-			competitionData.insert(new Competition(season, "Autodesk Oregon Regional", "OR"));
-			competitionData.insert(new Competition(season, "Chesapeake Regional", "MD"));
-			competitionData.insert(new Competition(season, "Chestnut Hill FIRST Robotics District Competition", "PHL"));
-			competitionData.insert(new Competition(season, "Finger Lakes Regional", "ROC"));
-			competitionData.insert(new Competition(season, "Greater Toronto East Regional", "ON"));
-			competitionData.insert(new Competition(season, "Lake Superior Regional", "DMN"));
-			competitionData.insert(new Competition(season, "Orlando Regional", "FL"));
-			competitionData.insert(new Competition(season, "Pittsburgh Regional", "PIT"));
-			competitionData.insert(new Competition(season, "Traverse City FIRST Robotics District Competition", "GT"));
-			competitionData.insert(new Competition(season, "Waterford FIRST Robotics District Competition", "OC1"));
-			competitionData.insert(new Competition(season, "WPI Regional", "WOR"));
-			competitionData.insert(new Competition(season, "Rutgers University FIRST Robotics District Competition", "NJ"));
-			competitionData.insert(new Competition(season, "Bayou Regional", "LA"));
-			competitionData.insert(new Competition(season, "Boilermaker Regional", "IN"));
-			competitionData.insert(new Competition(season, "Detroit FIRST Robotics District Competition", "DT"));
-			competitionData.insert(new Competition(season, "Festival de Robotique FRC a Montreal Regional", "QC"));
-			competitionData.insert(new Competition(season, "Los Angeles Regional", "CA"));
-			competitionData.insert(new Competition(season, "Peachtree Regional", "GA"));
-			competitionData.insert(new Competition(season, "Sacramento Regional", "SAC"));
-			competitionData.insert(new Competition(season, "Utah Regional co-sponsored by NASA and Platt", "UT"));
-			competitionData.insert(new Competition(season, "Virginia Regional", "VA"));
-			competitionData.insert(new Competition(season, "West Michigan FIRST Robotics District Competition", "MI"));
-			competitionData.insert(new Competition(season, "New York City Regional", "NY"));
-			competitionData.insert(new Competition(season, "Arizona Regional", "AZ"));
-			competitionData.insert(new Competition(season, "Boston Regional", "MA"));
-			competitionData.insert(new Competition(season, "Hawaii Regional sponsored by BAE Systems", "HI"));
-			competitionData.insert(new Competition(season, "Buckeye Regional", "OH"));
-			competitionData.insert(new Competition(season, "Colorado Regional", "CO"));
-			competitionData.insert(new Competition(season, "Lenape FIRST Robotics District Competition", "NJT"));
-			competitionData.insert(new Competition(season, "Midwest Regional", "IL"));
-			competitionData.insert(new Competition(season, "Niles FIRST Robotics District Competition", "SWM"));
-			competitionData.insert(new Competition(season, "Northville FIRST Robotics District Competition", "WCA"));
-			competitionData.insert(new Competition(season, "Palmetto Regional", "SC"));
-			competitionData.insert(new Competition(season, "Seattle Cascade Regional", "WA2"));
-			competitionData.insert(new Competition(season, "Seattle Olympic Regional", "WA"));
-			competitionData.insert(new Competition(season, "St. Louis Regional", "MO"));
-			competitionData.insert(new Competition(season, "Waterloo Regional", "WAT"));
-			competitionData.insert(new Competition(season, "Wisconsin Regional", "WI"));
-			competitionData.insert(new Competition(season, "Dallas East Regional sponsored by jcpenney", "DA"));
-			competitionData.insert(new Competition(season, "Dallas West Regional sponsored by jcpenney", "DA2"));
-			competitionData.insert(new Competition(season, "Greater Toronto West Regional", "ON2"));
-			competitionData.insert(new Competition(season, "Livonia FIRST Robotics District Competition", "WW"));
-			competitionData.insert(new Competition(season, "Minnesota 10000 Lakes Regional", "MN"));
-			competitionData.insert(new Competition(season, "Minnesota North Star Regional", "MN2"));
-			competitionData.insert(new Competition(season, "Mount Olive FIRST Robotics District Competition", "NJF"));
-			competitionData.insert(new Competition(season, "Northeast Utilities FIRST Connecticut Regional", "CT"));
-			competitionData.insert(new Competition(season, "Oklahoma Regional", "OK"));
-			competitionData.insert(new Competition(season, "SBPLI Long Island Regional", "LI"));
-			competitionData.insert(new Competition(season, "Silicon Valley Regional", "SJ"));
-			competitionData.insert(new Competition(season, "South Florida Regional", "SFL"));
-			competitionData.insert(new Competition(season, "Troy FIRST Robotics District Competition", "OC"));
-			competitionData.insert(new Competition(season, "Washington DC  Regional", "DC"));
-			competitionData.insert(new Competition(season, "Central Valley Regional", "CAF"));
-			competitionData.insert(new Competition(season, "Las Vegas Regional", "NV"));
-			competitionData.insert(new Competition(season, "Lone Star Regional", "TX"));
-			competitionData.insert(new Competition(season, "North Carolina Regional", "NC"));
-			competitionData.insert(new Competition(season, "Queen City Regional", "OHC"));
-			competitionData.insert(new Competition(season, "Spokane Regional", "WAS"));
-			competitionData.insert(new Competition(season, "Mid-Atlantic Robotics FRC Region Championship", "PA"));
-			competitionData.insert(new Competition(season, "Michigan FRC State Championship", "GL"));
-			competitionData.insert(new Competition(season, "FIRST Championship - Archimedes Division", "Archimedes"));
-			competitionData.insert(new Competition(season, "FIRST Championship - Curie Division", "Curie"));
-			competitionData.insert(new Competition(season, "FIRST Championship - Galileo Division", "Galileo"));
-			competitionData.insert(new Competition(season, "FIRST Championship - Newton Division", "Newton"));
-			competitionData.insert(new Competition(season, "FIRST Championship", "CMP"));
-		} catch (OurAllianceException e) {
-			e.printStackTrace();
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
-	}
-	
-	private void addTestData(Season season, Competition comp) {
-		SparseArray<String> teams = new SparseArray<String>();
-		teams.append(1676, "Pioneers");
-		teams.append(3637, "The Daleks");
-		teams.append(25, "Raider Robotix");
-		teams.append(75, "RoboRaiders");
-		teams.append(11, "Chief Delphi");
-		try {
-			for(int i=0;i<teams.size();++i) {
-				Team team = new Team(teams.keyAt(i), teams.valueAt(i));
-				team = teamData.insert(team);
-				teamScouting2013Data.insert(new TeamScouting2013(season, team));
-				competitionTeamData.insert(new CompetitionTeam(comp, team));
+	private boolean addCompetitions(Season season) {
+		comps = new HashMap<String,String>();
+    	increasePrimary();
+		putComp("STX", "Alamo Regional");
+		putComp("NH", "BAE Systems Granite State Regional");
+		putComp("KC", "Greater Kansas City Regional");
+		putComp("PAH", "Hatboro-Horsham FIRST Robotics District Competition");
+		putComp("GG", "Kettering University FIRST Robotics District Competition");
+		putComp("MIGL", "Gull Lake FIRST Robotics District Competition");
+		putComp("TN", "Smoky Mountains Regional");
+		putComp("SDC", "San Diego Regional");
+		putComp("IS", "Israel Regional");
+		putComp("OR", "Autodesk Oregon Regional");
+		putComp("MD", "Chesapeake Regional");
+		putComp("PHL", "Chestnut Hill FIRST Robotics District Competition");
+		putComp("ROC", "Finger Lakes Regional");
+		putComp("ON", "Greater Toronto East Regional");
+		putComp("DMN", "Lake Superior Regional");
+		putComp("FL", "Orlando Regional");
+		putComp("PIT", "Pittsburgh Regional");
+		putComp("GT", "Traverse City FIRST Robotics District Competition");
+		putComp("OC1", "Waterford FIRST Robotics District Competition");
+		putComp("WOR", "WPI Regional");
+		putComp("NJ", "Rutgers University FIRST Robotics District Competition");
+		putComp("LA", "Bayou Regional");
+		putComp("IN", "Boilermaker Regional");
+		putComp("DT", "Detroit FIRST Robotics District Competition");
+		putComp("QC", "Festival de Robotique FRC a Montreal Regional");
+		putComp("CA", "Los Angeles Regional");
+		putComp("GA", "Peachtree Regional");
+		putComp("SAC", "Sacramento Regional");
+		putComp("UT", "Utah Regional co-sponsored by NASA and Platt");
+		putComp("VA", "Virginia Regional");
+		putComp("MI", "West Michigan FIRST Robotics District Competition");
+		putComp("NY", "New York City Regional");
+		putComp("AZ", "Arizona Regional");
+		putComp("MA", "Boston Regional");
+		putComp("HI", "Hawaii Regional sponsored by BAE Systems");
+		putComp("OH", "Buckeye Regional");
+		putComp("CO", "Colorado Regional");
+		putComp("NJT", "Lenape FIRST Robotics District Competition");
+		putComp("IL", "Midwest Regional");
+		putComp("SWM", "Niles FIRST Robotics District Competition");
+		putComp("WCA", "Northville FIRST Robotics District Competition");
+		putComp("SC", "Palmetto Regional");
+		putComp("WA2", "Seattle Cascade Regional");
+		putComp("WA", "Seattle Olympic Regional");
+		putComp("MO", "St. Louis Regional");
+		putComp("WAT", "Waterloo Regional");
+		putComp("WI", "Wisconsin Regional");
+		putComp("DA", "Dallas East Regional sponsored by jcpenney");
+		putComp("DA2", "Dallas West Regional sponsored by jcpenney");
+		putComp("ON2", "Greater Toronto West Regional");
+		putComp("WW", "Livonia FIRST Robotics District Competition");
+		putComp("MN", "Minnesota 10000 Lakes Regional");
+		putComp("MN2", "Minnesota North Star Regional");
+		putComp("NJF", "Mount Olive FIRST Robotics District Competition");
+		putComp("CT", "Northeast Utilities FIRST Connecticut Regional");
+		putComp("OK", "Oklahoma Regional");
+		putComp("LI", "SBPLI Long Island Regional");
+		putComp("SJ", "Silicon Valley Regional");
+		putComp("SFL", "South Florida Regional");
+		putComp("OC", "Troy FIRST Robotics District Competition");
+		putComp("DC", "Washington DC  Regional");
+		putComp("CAF", "Central Valley Regional");
+		putComp("NV", "Las Vegas Regional");
+		putComp("TX", "Lone Star Regional");
+		putComp("NC", "North Carolina Regional");
+		putComp("OHC", "Queen City Regional");
+		putComp("WAS", "Spokane Regional");
+		putComp("PA", "Mid-Atlantic Robotics FRC Region Championship");
+		putComp("GL", "Michigan FRC State Championship");
+		putComp("Archimedes", "FIRST Championship - Archimedes Division");
+		putComp("Curie", "FIRST Championship - Curie Division");
+		putComp("Galileo", "FIRST Championship - Galileo Division");
+		putComp("Newton", "FIRST Championship - Newton Division");
+		putComp("CMP", "FIRST Championship");
+		Competition thisComp;
+		for (Map.Entry<String, String> each : comps.entrySet()) {
+			try {
+				thisComp = new Competition(season, each.getValue(), each.getKey());
+				Log.d(TAG,thisComp.toString());
+				competitionData.insert(thisComp);
+			} catch (OurAllianceException e) {
+				e.printStackTrace();
+				return false;
+			} catch (SQLException e) {
+				e.printStackTrace();
+				return false;
 			}
-			Team team = new Team(869, "Power Cord");
-			team = teamData.insert(team);
-			teamScouting2013Data.insert(new TeamScouting2013(season, team));
-			competitionTeamData.insert(new CompetitionTeam(comp, team, 1));
-		} catch (OurAllianceException e) {
-			e.printStackTrace();
-		} catch (SQLException e) {
-			e.printStackTrace();
+	    	increasePrimary();
 		}
+		return true;
+	}
+	
+	public void putComp(String code, String name) {
+		comps.put(code, name);
+    	increasePrimary();
+	}
+	
+	private void setFlag(int flag) {
+		this.flag = flag;
+        publishProgress(new Object[]{flag, primary, version+": "+status});
+	}
+	
+	private void increasePrimary() {
+        publishProgress(new Object[]{flag, ++primary, version+": "+status});
+	}
+	
+	private void increaseVersion() {
+        publishProgress(new Object[]{flag, primary, (++version)+": "+status});
+	}
+	
+	private void setStatus(CharSequence status) {
+		this.status = status;
+		Log.d(TAG, status.toString());
+        publishProgress(new Object[]{flag, primary, version+": "+status});
 	}
 }
